Only use cached outputs for the same input.
This changes updates the caching logic to ensure that a cached output
is only loaded if it was generated with the same inputs as the current
run. Previously, if a video was requested more than once for the same
Video ID, the previous output would be loaded regardless of whether
the request parameters or article content had changed since the last
run.

Change-Id: Ife88355637195911ac0db8ba41e03cdc9b5045ff